Moves HatchPattern to be a top level component.
