+++ src/components/DataTable/DataTable.less
@@ -18,14 +18,4 @@
 		text-align: center;
 		margin-bottom: @size-XS;
 	}
-
-	&-message-header {
-		height: @size-L;
-		width: 100%;
-	}
-
-	&-message-pattern {
-		stroke: @color-gray;
-		stroke-width: 1px;
-	}
 }
