+++ src/components/HatchPattern/HatchPattern.less
@@ -1,0 +1,11 @@
+.lucid-HatchPattern {
+	&-container {
+		height: @size-L;
+		width: 100%;
+	}
+
+	&-pattern {
+		stroke: @color-gray;
+		stroke-width: 1px;
+	}
+}
