+++ src/styles/components.less
@@ -66,3 +66,4 @@
 @import '../components/Overlay/Overlay';
 @import '../components/Dialog/Dialog';
 @import '../components/Autocomplete/Autocomplete'; // Should import after DropMenu
+@import '../components/HatchPattern/HatchPattern';
